pyproject.toml
[tool.poetry]
name = "flaskbs"
version = "0.1.0"
description = ""
authors = ["David Felce <davefelce@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.2"
python-dotenv = "^0.21.0"
flask-sqlalchemy = "^3.0.2"
flask-migrate = "^4.0.0"
psycopg2-binary = "^2.9.5"

[tool.poetry.group.dev.dependencies]
mypy = "^0.991"
black = "^22.12.0"
xenon = "^0.9.0"
pylint = "^2.15.9"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pylint-pytest = "^1.1.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
[tool.black]
line-length = 120

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 120
skip_gitignore = true
lines_between_types = 1

[tool.mypy]
disallow_untyped_defs = true
show_error_codes = true
warn_unused_ignores = true
ignore_missing_imports = true
disallow_incomplete_defs = true
warn_redundant_casts = true
# warn_unreachable = true
plugins = [
    "sqlalchemy.ext.mypy.plugin",
]

[tool.pylint.MASTER]
load-plugins=[
    "pylint_pytest",
    "pylint.extensions.mccabe"
]

[tool.pylint.messages_control]
max-line-length = 120
disable = [
    "logging-fstring-interpolation",
    "missing-module-docstring",
    "missing-function-docstring",
    "missing-class-docstring",
    "too-few-public-methods",
    "consider-using-f-string",
    "duplicate-code", # pylint definition of "duplicate" is way too loose (ex: fastapi route definitions are flagged).
    "invalid-name",
]
good-names=["k", "v", "i", "q", "fn", "ex", "dt", "tz"]
